Bug fixes on api call using retrofit client.  Displaying recipes in main layout.

+++ app/src/main/java/com/example/someone/bakingapp/models/IngredientModel.java
@@ -9,7 +9,7 @@ import com.google.gson.annotations.SerializedName;
 public class IngredientModel {
 
     @SerializedName("quantity")
-    public Integer quantity;
+    public float quantity;
 
     @SerializedName("measure")
     public String measure;
